Automated parsabilety of hex numbers ("0x...") (#35)
* Automated parsabilety of hex numbers starting with "0x"

When the NM client parses something like a Dictionary from JSON where input keys look like this:
"0x0000000000000000000000000000000000000000000000000000000000000000"
we currently fail. To fix this, we need to re-implement the dictionary parser as they do here (https://stackoverflow.com/a/7010231/1973207 )  with explicit `(UInt256)BigInteger.Parse(skey.Substring(2), NumberStyles.HexNumber);` in it. This would help automate this and allow the elimination of specialised dictionary parsers like that.